prepare ExecutionPayload for extension
