+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/versions/bellatrix/BeaconBlockBodyBuilderBellatrix.java
@@ -120,6 +120,6 @@ public class BeaconBlockBodyBuilderBellatrix extends BeaconBlockBodyBuilderAltai
                 deposits,
                 voluntaryExits,
                 syncAggregate,
-                payload));
+                payload.toVersionBellatrix().orElseThrow()));
   }
 }
