+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/versions/capella/BeaconBlockBodySchemaCapellaImpl.java
@@ -180,8 +180,8 @@ public class BeaconBlockBodySchemaCapellaImpl
   }
 
   @Override
-  public ExecutionPayloadSchema getExecutionPayloadSchema() {
-    return (ExecutionPayloadSchema) getFieldSchema9();
+  public ExecutionPayloadSchema<?> getExecutionPayloadSchema() {
+    return (ExecutionPayloadSchema<?>) getFieldSchema9();
   }
 
   @SuppressWarnings("unchecked")
