+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/ExecutionPayloadFields.java
@@ -31,7 +31,8 @@ public enum ExecutionPayloadFields implements SszFieldName {
   BASE_FEE_PER_GAS,
   BLOCK_HASH,
   TRANSACTIONS,
-  TRANSACTIONS_ROOT;
+  TRANSACTIONS_ROOT,
+  WITHDRAWALS_ROOT;
 
   private final String sszFieldName;
