+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/versions/bellatrix/ExecutionPayloadSchemaBellatrix.java
@@ -1,17 +1,4 @@
-/*
- * Copyright ConsenSys Software Inc., 2022
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.spec.datastructures.execution;
+package tech.pegasys.teku.spec.datastructures.execution.versions.bellatrix;
 
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.BASE_FEE_PER_GAS;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.BLOCK_HASH;
@@ -26,29 +13,35 @@ import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFi
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.RECEIPTS_ROOT;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.STATE_ROOT;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.TIMESTAMP;
-import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.TRANSACTIONS_ROOT;
+import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.TRANSACTIONS;
 
+import java.util.List;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.units.bigints.UInt256;
 import tech.pegasys.teku.infrastructure.bytes.Bytes20;
+import tech.pegasys.teku.infrastructure.ssz.SszList;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszByteList;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszByteVector;
 import tech.pegasys.teku.infrastructure.ssz.containers.ContainerSchema14;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszBytes32;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt256;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt64;
+import tech.pegasys.teku.infrastructure.ssz.schema.SszListSchema;
 import tech.pegasys.teku.infrastructure.ssz.schema.SszPrimitiveSchemas;
 import tech.pegasys.teku.infrastructure.ssz.schema.collections.SszByteListSchema;
 import tech.pegasys.teku.infrastructure.ssz.schema.collections.SszByteVectorSchema;
 import tech.pegasys.teku.infrastructure.ssz.tree.TreeNode;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.config.SpecConfigBellatrix;
-import tech.pegasys.teku.spec.datastructures.execution.versions.bellatrix.ExecutionPayloadSchemaBellatrix;
+import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
+import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadSchema;
+import tech.pegasys.teku.spec.datastructures.execution.Transaction;
+import tech.pegasys.teku.spec.datastructures.execution.TransactionSchema;
 
-public class ExecutionPayloadHeaderSchema
+public class ExecutionPayloadSchemaBellatrix
     extends ContainerSchema14<
-        ExecutionPayloadHeader,
+        ExecutionPayloadBellatrix,
         SszBytes32,
         SszByteVector,
         SszBytes32,
@@ -62,14 +55,14 @@ public class ExecutionPayloadHeaderSchema
         SszByteList,
         SszUInt256,
         SszBytes32,
-        SszBytes32> {
+        SszList<Transaction>>
+    implements ExecutionPayloadSchema<ExecutionPayloadBellatrix> {
 
-  private final ExecutionPayloadHeader defaultExecutionPayloadHeader;
-  private final ExecutionPayloadHeader executionPayloadHeaderOfDefaultPayload;
+  private final ExecutionPayloadBellatrix defaultExecutionPayload;
 
-  public ExecutionPayloadHeaderSchema(final SpecConfigBellatrix specConfig) {
+  public ExecutionPayloadSchemaBellatrix(final SpecConfigBellatrix specConfig) {
     super(
-        "ExecutionPayloadHeader",
+        "ExecutionPayloadBellatrix",
         namedSchema(PARENT_HASH, SszPrimitiveSchemas.BYTES32_SCHEMA),
         namedSchema(FEE_RECIPIENT, SszByteVectorSchema.create(Bytes20.SIZE)),
         namedSchema(STATE_ROOT, SszPrimitiveSchemas.BYTES32_SCHEMA),
@@ -83,27 +76,15 @@ public class ExecutionPayloadHeaderSchema
         namedSchema(EXTRA_DATA, SszByteListSchema.create(specConfig.getMaxExtraDataBytes())),
         namedSchema(BASE_FEE_PER_GAS, SszPrimitiveSchemas.UINT256_SCHEMA),
         namedSchema(BLOCK_HASH, SszPrimitiveSchemas.BYTES32_SCHEMA),
-        namedSchema(TRANSACTIONS_ROOT, SszPrimitiveSchemas.BYTES32_SCHEMA));
+        namedSchema(
+            TRANSACTIONS,
+            SszListSchema.create(
+                new TransactionSchema(specConfig), specConfig.getMaxTransactionsPerPayload())));
 
-    final ExecutionPayload defaultExecutionPayload =
-        new ExecutionPayloadSchemaBellatrix(specConfig).getDefault();
-
-    this.executionPayloadHeaderOfDefaultPayload =
-        createFromExecutionPayload(defaultExecutionPayload);
-
-    this.defaultExecutionPayloadHeader = createFromBackingNode(getDefaultTree());
+    this.defaultExecutionPayload = createFromBackingNode(getDefaultTree());
   }
 
-  public SszByteListSchema<?> getExtraDataSchema() {
-    return (SszByteListSchema<?>) getFieldSchema10();
-  }
-
-  @Override
-  public ExecutionPayloadHeader createFromBackingNode(TreeNode node) {
-    return new ExecutionPayloadHeader(this, node);
-  }
-
-  public ExecutionPayloadHeader create(
+  public ExecutionPayload create(
       Bytes32 parentHash,
       Bytes20 feeRecipient,
       Bytes32 stateRoot,
@@ -117,8 +98,8 @@ public class ExecutionPayloadHeaderSchema
       Bytes extraData,
       UInt256 baseFeePerGas,
       Bytes32 blockHash,
-      Bytes32 transactionsRoot) {
-    return new ExecutionPayloadHeader(
+      List<Bytes> transactions) {
+    return new ExecutionPayloadBellatrix(
         this,
         SszBytes32.of(parentHash),
         SszByteVector.fromBytes(feeRecipient.getWrappedBytes()),
@@ -133,39 +114,35 @@ public class ExecutionPayloadHeaderSchema
         getExtraDataSchema().fromBytes(extraData),
         SszUInt256.of(baseFeePerGas),
         SszBytes32.of(blockHash),
-        SszBytes32.of(transactionsRoot));
-  }
-
-  public ExecutionPayloadHeader createFromExecutionPayload(
-      final ExecutionPayload executionPayload) {
-    return new ExecutionPayloadHeader(
-        this,
-        SszBytes32.of(executionPayload.getParentHash()),
-        SszByteVector.fromBytes(executionPayload.getFeeRecipient().getWrappedBytes()),
-        SszBytes32.of(executionPayload.getStateRoot()),
-        SszBytes32.of(executionPayload.getReceiptsRoot()),
-        SszByteVector.fromBytes(executionPayload.getLogsBloom()),
-        SszBytes32.of(executionPayload.getPrevRandao()),
-        SszUInt64.of(executionPayload.getBlockNumber()),
-        SszUInt64.of(executionPayload.getGasLimit()),
-        SszUInt64.of(executionPayload.getGasUsed()),
-        SszUInt64.of(executionPayload.getTimestamp()),
-        getExtraDataSchema().fromBytes(executionPayload.getExtraData()),
-        SszUInt256.of(executionPayload.getBaseFeePerGas()),
-        SszBytes32.of(executionPayload.getBlockHash()),
-        SszBytes32.of(executionPayload.getTransactions().hashTreeRoot()));
+        transactions.stream()
+            .map(getTransactionSchema()::fromBytes)
+            .collect(getTransactionsSchema().collector()));
   }
 
   @Override
-  public ExecutionPayloadHeader getDefault() {
-    return defaultExecutionPayloadHeader;
+  public TransactionSchema getTransactionSchema() {
+    return (TransactionSchema) getTransactionsSchema().getElementSchema();
   }
 
-  public ExecutionPayloadHeader getHeaderOfDefaultPayload() {
-    return executionPayloadHeaderOfDefaultPayload;
+  @SuppressWarnings("unchecked")
+  @Override
+  public SszListSchema<Transaction, ?> getTransactionsSchema() {
+    return (SszListSchema<Transaction, ?>) getFieldSchema13();
   }
 
-  public long getBlindedNodeGeneralizedIndex() {
-    return getChildGeneralizedIndex(getFieldIndex(TRANSACTIONS_ROOT));
+  @SuppressWarnings("unchecked")
+  @Override
+  public SszByteListSchema<?> getExtraDataSchema() {
+    return (SszByteListSchema<?>) getFieldSchema10();
+  }
+
+  @Override
+  public ExecutionPayloadBellatrix getDefault() {
+    return defaultExecutionPayload;
+  }
+
+  @Override
+  public ExecutionPayloadBellatrix createFromBackingNode(TreeNode node) {
+    return new ExecutionPayloadBellatrix(this, node);
   }
 }
