+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/schemas/SchemaDefinitionsBellatrix.java
@@ -27,7 +27,7 @@ import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.bellatrix
 import tech.pegasys.teku.spec.datastructures.builder.BuilderBidSchema;
 import tech.pegasys.teku.spec.datastructures.builder.SignedBuilderBidSchema;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadHeaderSchema;
-import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadSchema;
+import tech.pegasys.teku.spec.datastructures.execution.versions.bellatrix.ExecutionPayloadSchemaBellatrix;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateSchema;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.bellatrix.BeaconStateBellatrix;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.bellatrix.BeaconStateSchemaBellatrix;
@@ -111,8 +111,8 @@ public class SchemaDefinitionsBellatrix extends SchemaDefinitionsAltair {
     return signedBlindedBeaconBlockSchema;
   }
 
-  public ExecutionPayloadSchema getExecutionPayloadSchema() {
-    return beaconBlockBodySchema.getExecutionPayloadSchema();
+  public ExecutionPayloadSchemaBellatrix getExecutionPayloadSchema() {
+    return (ExecutionPayloadSchemaBellatrix) beaconBlockBodySchema.getExecutionPayloadSchema();
   }
 
   public ExecutionPayloadHeaderSchema getExecutionPayloadHeaderSchema() {
